Переделан механизм создания книг, разные книги теперь относятся к разным классам. Исправлены мелкие баги.
