+++ class_items.py
@@ -542,192 +542,3 @@ class Key:
         
         tprint(
             self.game, f'{who_is_using.name} не знает, как можно использовать ключ если не открывать им что-то.')
-
-
-class Book:
-    
-    _names = {
-        "nom": "книга",
-        "accus": "книгу",
-        "gen": "книги",
-        "dat": "книге",
-        "prep": "книге",
-        "inst": "книгой"
-      }
-    
-    _descriptions = (
-        {
-            "nom": "Старая",
-            "accus": "Старую",
-            "gen": "Старой",
-            "dat": "Старой",
-            "prep": "Старой",
-            "inst": "Старой"
-          },
-          {
-            "nom": "Древняя",
-            "accus": "Древнюю",
-            "gen": "Древней",
-            "dat": "Древней",
-            "prep": "Древней",
-            "inst": "Древней"
-          },
-          {
-            "nom": "Пыльная",
-            "accus": "Пыльную",
-            "gen": "Пыльной",
-            "dat": "Пыльной",
-            "prep": "Пыльной",
-            "inst": "Пыльной"
-          },
-          {
-            "nom": "Зачитанная",
-            "accus": "Зачитанную",
-            "gen": "Зачитанной",
-            "dat": "Зачитанной",
-            "prep": "Зачитанной",
-            "inst": "Зачитанной"
-          },
-          {
-            "nom": "Новая",
-            "accus": "Новую",
-            "gen": "Новой",
-            "dat": "Новой",
-            "prep": "Новой",
-            "inst": "Новой"
-          },
-          {
-            "nom": "Потрепанная",
-            "accus": "Потрепанную",
-            "gen": "Потрепанной",
-            "dat": "Потрепанной",
-            "prep": "Потрепанной",
-            "inst": "Потрепанной"
-          },
-          {
-            "nom": "Красивая",
-            "accus": "Красивую",
-            "gen": "Красивой",
-            "dat": "Красивой",
-            "prep": "Красивой",
-            "inst": "Красивой"
-          },
-          {
-            "nom": "Большая",
-            "accus": "Большую",
-            "gen": "Большой",
-            "dat": "Большой",
-            "prep": "Большой",
-            "inst": "Большой"
-          }
-    )
-    
-    _texts = (
-        [0, "Книга повествует о временах мастеров-фехтовальщиков. Тогда величайшим достижением считалось заколоть соперника в честном поединке на шпагах."],
-        [0, "В книге во всех красках описывается история двух друзей-авантюристов, которые берутся за любое, даже самое опасное дело. Попав в беду они всегда чудом избегают гибели благодаря их умениям в бою на шпагах."],
-        [0, "Это история о молодой девушке, отец которой считает, что она добьется всего сама, если будет обладать характером и умением фехтовать. Уроки отца девушки описаны подробно и основательно, что позволяет читателю почерпнуть много нового."],
-        [1, "Книга рассказывает о злобных пиратах под предводительством знаменитого капитана Рыжая (с проседью) Борода, который был великим мастером сабли."],
-        [1, "Повествование в книге ведется от лица боцмана пиратской шхуны, бороздившей южные моря. Команда корабля не гнушалась любой добычи и нападала на любые проходящие мимо суда. Не сгинуть в морской пучине главному герою помогает его мастерское владение саблей."],
-        [1, "Автор книги провел несколько лет в пиратском логове в качестве врача. Наблюдая за жизнью пиратов, за их тренировками, он составил подробное описание боевого стиля пиратских воинов."],
-        [2, "В книге описывается жизнь вождя викингов Олафа, который ходил в походы на соседние земли и прославился тем, что дрался только своей любимой дубиной."],
-        [2, "Книга представляет собой вольное изложение легенд северных народов с особым упором на ратные подвиги героев и их умения в применении ударного оружия."],
-        [2, "История кузница в норвежской деревне начинается мирно и размеренно, но резко превращается в абсолютный кошмар после эпизода убийства его семьи воинами соседнего племени. Бросив кузницу и вооружившись лишь своим молотом герой отправляется мстить."],
-        [3, "Книга рассказывает про команду взломщиков сейфов, которая сталкивается с самым сложным вызовом в своей карьере, пытаясь проникнуть в неприступное хранилище, охраняемое страшными ловушками."],
-        [3, "Книга описывает мир, где заклинания могут открывать любые двери. Группа храбрых плутов бросает вызов забытому замку, полному магических загадок и смертельных ловушек, чтобы обнаружить сокровище, способное изменить судьбы."],
-        [3, "В книге описывается как в зачарованном лесу, где каждое дерево и камень хранит свои тайны, молодой вор предпринимает опасное путешествие к сердцу древнего лабиринта, полного мистических ловушек и неизведанных опасностей, в поисках затерянного сокровища великих магов."]
-    )
-    
-    _decorations = (
-        "о мушкетерах",
-        "о пиратах",
-        "о викингах",
-        "о взломщиках"
-    )
-    
-    _weapon_types = (
-        "колющее",
-        "рубящее",
-        "ударное",
-        0
-    )
-    
-    _base_price = 7
-    
-    _base_price_die = [8]
-    
-    def __init__(self, game):
-        self.game = game
-        self.empty = False
-        self.type = roll([4]) - 1
-        self.create_description()
-        self.name = self.lexemes['nom']
-        available_texts = [i for i in Book._texts if i[0] == self.type]
-        self.text = randomitem(available_texts)
-        self.weapon_type = Book._weapon_types[self.type]
-        self.can_use_in_fight = False
-        self.base_price = self.define_base_price()
-
-
-    def __format__(self, format:str) -> str:
-        return self.lexemes.get(format, '')
-
-    
-    def __str__(self):
-        return self.name
-    
-    
-    def check_name(self, message:str) -> bool:
-        names_list = self.get_names_list(['nom', "accus"])
-        return message.lower() in names_list
-    
-    
-    def get_names_list(self, cases:list=None) -> list:
-        names_list = ['книга', 'книгу', 'книжка', 'книжку']
-        for case in cases:
-            names_list.append(self.lexemes.get(case, '').lower())
-        return names_list
-            
-    
-    def define_base_price(self) -> int:
-        return Book._base_price + roll(Book._base_price_die)
-               
-   
-    def create_description(self):
-        description_dict = randomitem(Book._descriptions)
-        name_dict = Book._names
-        self.lexemes = {}
-        for lexeme in name_dict:
-            self.lexemes[lexeme] =f'{description_dict[lexeme]} {name_dict[lexeme]} {Book._decorations[self.type]}'
-        return True
-        
-    
-    def get_mastery_string(self, who):
-        return f'{who.name} теперь немного лучше знает, как использовать {self.weapon_type} оружие.'
-    
-    
-    def place(self, floor, room=None):
-        if not room:
-            room = floor.get_random_room_with_furniture()
-        furniture = randomitem(room.furniture)
-        furniture.put(self)
-        return True
-    
-    
-    def show(self):
-        return self.lexemes['nom']
-    
-    
-    def use(self, who_using, in_action:bool=False):
-        if in_action:
-            tprint(self.game, 'Сейчас абсолютно не подходящее время для чтения.')
-            return False
-        else:
-            return who_using.read(self)
-    
-    
-    def take(self, who):
-        if not who.backpack.no_backpack:
-            who.backpack.append(self)
-            tprint(self.game, f'{who.name} забирает {self:accus} себе.')
-            return True
-        return False
